feat: removed unnecesary abstract class
