+++ src/contracts/account.ts
@@ -2,7 +2,7 @@ import { Account, Connection } from '@near-js/accounts'
 import { KeyPair } from '@near-js/crypto'
 import { InMemoryKeyStore } from '@near-js/keystores'
 
-import { DONT_CARE_ACCOUNT_ID } from '@contracts/near/constants'
+import { DONT_CARE_ACCOUNT_ID } from '@contracts/constants'
 
 type SetConnectionArgs =
   | {
