+++ src/contracts/index.ts
@@ -1,3 +1,11 @@
-export { ChainSignatureContract, type SignArgs } from './ChainSignatureContract'
+// TODO: fix signAndSend
+// import * as signAndSend from './signAndSend'
+import * as transaction from './transaction'
+export * from './ChainSignatureContract'
 
-export * as near from './near'
+const utils = {
+  transaction,
+  // signAndSend,
+}
+
+export { utils }
