--- src/contracts/near/index.ts
@@ -1,10 +1,0 @@
-import * as signAndSend from './signAndSend'
-import * as transaction from './transaction'
-export * from './ChainSignatureContract'
-
-const utils = {
-  transaction,
-  signAndSend,
-}
-
-export { utils }
